Include hash of weights in the key of the primitive cache for aarch64 lowp gemm
This goes in hand with oneapi-src/oneDNN#2194 which sets the weights
as constant for Arm Compute Library (ACL) lowp gemm objects.

This is a temp fix which is needed because the lowp gemm ACL object
in oneDNN is now holding  more state. Hence, we need to make sure that
the cahce maps a layer with a specific set of weights to the oneDNN
primitive storing those weights.

We're currently working on the proper fix for this which involves
making lowp_gemm stateless and fixed-format in ACL and oneDNN.